==> Code/04_h2_fec_prediction.Rmd <==
---
title: "h2 fecundity analysis"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      cache = FALSE)

library(tidyverse)
library(cowplot)
library(readxl)
library(cvTools)
library(latex2exp)

source("ggplot_theme.R")
```

## Regression on handcounted images

```{r}
linear_threshold <- 53

# Load data for handcounts
HC <- suppressWarnings(
  read_excel("../Data/feclife_with-image-ids.xlsx") %>% 
    dplyr::select(cameraid, handcount, training_set) %>% 
    filter(training_set == "yes")
)

load("../Data/area_summation.Rda")

areas <- areas %>% 
  filter(lower_thresh == linear_threshold)

# Join handcounts with areas
HC <- inner_join(HC, areas, by = "cameraid") %>% 
  as.data.frame()

# Load image dimensions and merge
img_size <- read_csv("../Data/image_dimensions.csv")
HC <- left_join(HC, img_size, by = "cameraid")
```

Training image size

```{r}
HC %>% 
  ggplot(aes(img_size)) +
  geom_histogram()
```

## Prediction for all images

Linear model predicting egg count from area:

```{r fm_linear, results="hide", warning=FALSE}
fm_lm1 <- lm(handcount ~ I(area^0.5) - 1, data = HC)
fm_lm2 <- lm(handcount ~ I(area^0.5) + img_size - 1, data = HC)
fm_lm3 <- lm(handcount ~ I(area^0.5) * img_size - 1, data = HC)
```

Cross validation

```{r}
set.seed(4357896)

cv_fm1 <- cvFit(fm_lm1, K = 10, R = 1000, data = HC, y = HC$handcount)
cv_fm2 <- cvFit(fm_lm2, K = 10, R = 1000, data = HC, y = HC$handcount)
cv_fm3 <- cvFit(fm_lm3, K = 10, R = 1000, data = HC, y = HC$handcount)
```

```{r}
cvSelect(cv_fm1, cv_fm2, cv_fm3)
anova(fm_lm1, fm_lm2, fm_lm3, test = "Chisq")
```

```{r}
area <- seq(0, 1.25 * max(HC$area),
            length.out = 200)
img_size <- min(HC$img_size):max(HC$img_size)
new_data <- crossing(area, img_size) %>% 
  filter(img_size %% 250 == 0)

new_data$lm_pred <- predict(fm_lm2, newdata = new_data)

ggplot() +
  geom_line(data = new_data, aes(x = sqrt(area), y = lm_pred,
                                 color = factor(img_size))) +
  geom_point(data = HC, aes(x = sqrt(area), y = handcount))
```

## Evaluation of linear model

```{r}
HC$lm_pred <- predict(fm_lm2)

cor(HC$handcount, HC$lm_pred)

mean(abs(HC$handcount - HC$lm_pred))

HC_no_zero <- HC %>% filter(handcount > 50 & handcount < 350)
mean(abs(HC_no_zero$lm_pred - HC_no_zero$handcount) / HC_no_zero$handcount)

HC <- HC %>% 
  mutate(d = abs(lm_pred - handcount))

HC %>% 
  # filter(handcount > 50 & handcount < 350) %>% 
  ggplot(aes(y = d, x = handcount)) +
  geom_point() +
  geom_smooth(method = "lm")

summary(lm(d ~ handcount, HC))
```

Training set images with predicted negative counts

```{r}
HC %>% filter(lm_pred < 0) %>% arrange(lm_pred)
```

```{r}
HC %>% ggplot(aes(x = handcount, y = lm_pred)) +
  geom_abline(slope = 1) +
  geom_point()
```

## Predict for whole data set

```{r}
suppressWarnings(
  M <- read_excel("../Data/feclife_with-image-ids.xlsx") %>% 
    mutate(drop_from_lifespan = if_else(is.na(drop_from_lifespan),
                                        "no", drop_from_lifespan))
)

# Load areas
load("../Data/area_summation.Rda")
areas <- areas %>% 
  filter(lower_thresh == linear_threshold)

M <- left_join(M, areas, by = "cameraid") %>% 
  filter(drop_from_lifespan == "no")

# Load image dimensions and merge
img_size <- read_csv("../Data/image_dimensions.csv")
M <- left_join(M, img_size, by = "cameraid") %>% 
  select(-lower_thresh, -drop_from_lifespan)

M$predicted_count_linear <- predict(fm_lm2, newdata = M)
```

Prediction image size

```{r}
img_size %>% 
  ggplot(aes(img_size)) +
  geom_histogram()
```

## Some exploration

```{r}
ggplot(M, aes(predicted_count_linear)) +
  geom_histogram(bins = 30) +
  labs(x = "Predicted egg count", y = "count")
```

Check high and low predicted eggs

```{r}
M %>% 
  select(cameraid, predicted_count_linear) %>% 
  arrange(desc(predicted_count_linear))
```

```{r}
M %>% 
  select(cameraid, predicted_count_linear) %>% 
  arrange(predicted_count_linear)
```

# Save output for further filtering & analysis

```{r}
save(M, file = "../Data/predicted_egg_counts.rda")
load("../Data/predicted_egg_counts.rda")
```

# Visualize performance of model

```{r}
HC$grp <- seq(1, nrow(HC))

HG <- gather(HC[ , c(2, 5, 6, 7, 8)],
             key = "key", value = "egg_count", 
             c(handcount, lm_pred)) %>% 
  mutate(key = if_else(key == "handcount", "Handcounted", "Predicted"))

p1 <- HG %>% 
  ggplot(aes(y = egg_count, x = sqrt(area), color = key, group = grp)) +
  geom_point(alpha = 1/3) +
  geom_line(alpha = 1/3) +
  scale_color_manual(values = c("darkred", "blue"), name = "") +
  labs(x = TeX("Image Area ($\\sqrt{pixels}$)"), y = "Egg Count") +
  theme(legend.position = c(0.05, 0.9)) +
  my_theme
p1

save(p1, file = "../Figures/egg_count_area.Rda")

# Replace negative values with 0
M$predicted_count_linear[M$predicted_count_linear < 0] <- 0

p2 <- M %>%
  ggplot(aes(predicted_count_linear)) +
  geom_histogram(bins = 30) +
  labs(x = "Predicted Egg Count", y = "Count") +
  my_theme
save(p2, file = "../Figures/predicted_egg_count.Rda")
```
